requirements.txt
Package                    Version    
-------------------------- -----------
asn1crypto                 1.5.1      
certifi                    2022.5.18.1
cffi                       1.15.0     
charset-normalizer         2.0.12     
click                      8.1.3      
cryptography               36.0.2     
DateTime                   4.4        
Flask                      2.0.2      
idna                       3.3        
itsdangerous               2.1.2      
Jinja2                     3.1.2      
MarkupSafe                 2.1.1      
numpy                      1.22.4     
oscrypto                   1.3.0      
pandas                     1.4.2      
pip                        20.0.2     
pkg-resources              0.0.0      
pycparser                  2.21       
pycryptodomex              3.14.1     
PyJWT                      2.4.0      
pyOpenSSL                  21.0.0     
python-dateutil            2.8.2      
pytz                       2022.1     
requests                   2.27.1     
setuptools                 44.0.0     
six                        1.16.0     
snowflake-connector-python 2.7.7      
urllib3                    1.26.9     
Werkzeug                   2.1.2      
zope.interface             5.4.0      
